Part1/utils/generate_token.py
from jose import jwt
from datetime import datetime, timedelta

SECRET_KEY = "your_secret_key"
ALGORITHM = "HS256"

def generate_token():
    payload = {
        "sub": "service_center_001",
        "role": "service_center",
        "exp": datetime.utcnow() + timedelta(hours=6)
    }
    return jwt.encode(payload, SECRET_KEY, algorithm=ALGORITHM)

print(generate_token())
