requirements.txt
fastapi
uvicorn
pymongo
python-jose[cryptography]
geopy
pandas
pyodbc
sqlalchemy
langchain
langchain-community
langchain-experimental
langchain-google-genai
google-generativeai
streamlit
pandas
xgboost
scikit-learn
joblib
langchain
langchain-community
langchain-google-genai
chromadb
faiss-cpu
pypdf
python-dotenv
streamlit


